+++ vitepress-theme-teeker/src/config/index.ts
@@ -144,7 +144,7 @@ export default function tkThemeConfig(config: TkThemeConfig & UserConfig<Default
       // 解决项目启动后终端打印 Scss 的废弃警告：The legacy JS API is deprecated and will be removed in Dart Sass 2.0.0.
       css: { preprocessorOptions: { scss: { api: "modern" } } },
       optimizeDeps: {
-        include: ["element-plus", "@giscus/vue", "@waline/client"],
+        include: ["element-plus","@element-plus/icons-vue", "@giscus/vue", "@waline/client"],
       },
     },
     markdown: {