+++ plugins/vitepress-plugin-file-content-loader/CHANGELOG.md
@@ -1,5 +1,17 @@
 # vitepress-plugin-file-content-loader
 
+## 1.0.5
+
+### Patch Changes
+
+- 修复版本输出不正确问题
+
+## 1.0.4
+
+### Patch Changes
+
+- 新增日志功能，优化日志输出格式
+
 ## 1.0.3
 
 ### Patch Changes